+++ src/command/docs/open.rs
@@ -1,8 +1,7 @@
-use crate::{anyhow, command::RoverOutput, Result};
+use crate::{anyhow, command::RoverOutput, utils::color::Style, Result};
 
 use super::shortlinks;
 
-use ansi_term::Colour::{Cyan, Yellow};
 use saucer::{clap, Parser};
 use serde::Serialize;
 
@@ -17,13 +16,13 @@ pub struct Open {
 impl Open {
     pub fn run(&self) -> Result<RoverOutput> {
         let url = shortlinks::get_url_from_slug(&self.slug);
-        let yellow_browser_var = format!("{}", Yellow.normal().paint("$BROWSER"));
-        let cyan_url = format!("{}", Cyan.normal().paint(&url));
+        let painted_browser_var = Style::Command.paint("$BROWSER");
+        let painted_url = Style::Link.paint(&url);
 
         if let Some(browser_override) = std::env::var_os("BROWSER") {
             eprintln!(
                 "Opening {} with the application specified by {}.",
-                &cyan_url, &yellow_browser_var
+                &painted_url, &painted_browser_var
             );
             if let Err(e) = Command::new(&browser_override).arg(&url).status() {
                 Err(anyhow!(
@@ -35,7 +34,7 @@ impl Open {
                 Ok(())
             }
         } else {
-            eprintln!("Opening {} with your default browser. This can be overridden by setting the {} environment variable.", &cyan_url, &yellow_browser_var);
+            eprintln!("Opening {} with your default browser. This can be overridden by setting the {} environment variable.", &painted_url, &painted_browser_var);
             opener::open(&url)?;
             Ok(())
         }?;