+++ src/command/graph/fetch.rs
@@ -1,4 +1,3 @@
-use ansi_term::Colour::{Cyan, Yellow};
 use saucer::{clap, Parser};
 use serde::Serialize;
 
@@ -7,6 +6,7 @@ use rover_client::operations::graph::fetch::{self, GraphFetchInput};
 use crate::command::RoverOutput;
 use crate::options::{GraphRefOpt, ProfileOpt};
 use crate::utils::client::StudioClientConfig;
+use crate::utils::color::Style;
 use crate::Result;
 
 #[derive(Debug, Serialize, Parser)]
@@ -24,8 +24,8 @@ impl Fetch {
         let graph_ref = self.graph.graph_ref.to_string();
         eprintln!(
             "Fetching SDL from {} using credentials from the {} profile.",
-            Cyan.normal().paint(&graph_ref),
-            Yellow.normal().paint(&self.profile.profile_name)
+            Style::Link.paint(&graph_ref),
+            Style::Command.paint(&self.profile.profile_name)
         );
 
         let fetch_response = fetch::run(