+++ xtask/src/tools/mod.rs
@@ -1,11 +1,14 @@
 mod cargo;
 mod git;
+mod lychee;
 mod make;
 mod npm;
 mod runner;
 
 pub(crate) use cargo::CargoRunner;
 pub(crate) use git::GitRunner;
+#[cfg(not(windows))]
+pub(crate) use lychee::LycheeRunner;
 pub(crate) use make::MakeRunner;
 pub(crate) use npm::NpmRunner;
 pub(crate) use runner::Runner;